Add new workflows

+++ run.sh
@@ -1,8 +1,10 @@
 #!/usr/bin/bash
 
 # install long term support node.js
-source "$HOME"/.config/nvm/nvm.sh
-nvm install --lts
+if [[ -f "$HOME"/.config/nvm/nvm.sh ]]; then
+	source "$HOME"/.config/nvm/nvm.sh
+	nvm install --lts
+fi
 
 # compile the smart contracts
 npx hardhat compile
